+++ SQL/0000-00-01-Permission.sql
@@ -106,7 +106,9 @@ INSERT INTO `permissions` VALUES
     (58,'electrophysiology_browser_view_allsites', 'View all-sites Electrophysiology Browser pages', 2),
     (59,'electrophysiology_browser_view_site', 'View own site Electrophysiology Browser pages', 2),
     (60,'battery_manager_view','View Battery Manager',2),
-    (61,'battery_manager_edit','Add, activate, and deactivate entries in Test Battery',2);
+    (61,'battery_manager_edit','Add, activate, and deactivate entries in Test Battery',2),
+    (62,'module_manager_view', 'Module Manager: access the module', 2),
+    (63,'module_manager_edit', 'Module Manager: edit installed modules', 2);
 
 
 INSERT INTO `user_perm_rel` (userID, permID)