+++ webpack.config.js
@@ -59,6 +59,7 @@ const config = [{
     './modules/imaging_uploader/js/index.js': './modules/imaging_uploader/jsx/index.js',
     './modules/acknowledgements/js/acknowledgementsIndex.js': './modules/acknowledgements/jsx/acknowledgementsIndex.js',
     './modules/new_profile/js/NewProfileIndex.js': './modules/new_profile/jsx/NewProfileIndex.js',
+    './modules/module_manager/js/modulemanager.js': './modules/module_manager/jsx/modulemanager.js',
     './modules/imaging_qc/js/imagingQCIndex.js': './modules/imaging_qc/jsx/imagingQCIndex.js',
     './modules/server_processes_manager/js/server_processes_managerIndex.js': './modules/server_processes_manager/jsx/server_processes_managerIndex.js',
     './modules/document_repository/js/docIndex.js': './modules/document_repository/jsx/docIndex.js',